refactor(builder): split bundle sender from builder server
