+++ src/builder/task.rs
@@ -1,10 +1,18 @@
-use std::{collections::HashMap, sync::Arc, time::Duration};
+use std::{
+    collections::HashMap,
+    sync::{atomic::AtomicBool, Arc},
+    time::Duration,
+};
 
 use anyhow::{bail, Context};
 use ethers::types::{Address, H256};
 use ethers_signers::Signer;
 use rusoto_core::Region;
-use tokio::{select, sync::broadcast, time};
+use tokio::{
+    select,
+    sync::{broadcast, mpsc},
+    time,
+};
 use tokio_util::sync::CancellationToken;
 use tonic::{
     async_trait,
@@ -16,6 +24,7 @@ use crate::{
     builder::{
         self,
         bundle_proposer::{self, BundleProposerImpl},
+        bundle_sender::BundleSender,
         emit::BuilderEvent,
         sender::get_sender,
         server::{BuilderImpl, DummyBuilder},
@@ -158,11 +167,17 @@ impl Task for BuilderTask {
             tracker_settings,
         )
         .await?;
-        let builder_settings = builder::server::Settings {
+
+        let builder_settings = builder::bundle_sender::Settings {
             replacement_fee_percent_increase: self.args.replacement_fee_percent_increase,
             max_fee_increases: self.args.max_fee_increases,
         };
-        let builder = Arc::new(BuilderImpl::new(
+        let manual_bundling_mode = Arc::new(AtomicBool::new(false));
+        let (send_bundle_tx, send_bundle_rx) = mpsc::channel(1);
+
+        let bundle_sender = BundleSender::new(
+            manual_bundling_mode.clone(),
+            send_bundle_rx,
             self.args.chain_id,
             beneficiary,
             self.args.eth_poll_interval,
@@ -173,14 +188,14 @@ impl Task for BuilderTask {
             provider,
             builder_settings,
             self.event_sender.clone(),
-        ));
+        );
 
-        let _builder_loop_guard = {
-            let builder = Arc::clone(&builder);
-            SpawnGuard::spawn_with_guard(async move { builder.send_bundles_in_loop().await })
+        let _sender_loop_guard = {
+            SpawnGuard::spawn_with_guard(async move { bundle_sender.send_bundles_in_loop().await })
         };
 
         // gRPC server
+        let builder = BuilderImpl::new(manual_bundling_mode, send_bundle_tx);
         let builder_server = BuilderServer::new(builder);
 
         let reflection_service = tonic_reflection::server::Builder::configure()
